backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
structlog>=24.0.0
httpx>=0.27.0
fastmcp>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
